+++ src/apps/abracadabra/arbitrum/abracadabra.arbitrum.constants.ts
@@ -1,3 +1,5 @@
+import { HOURS_PER_YEAR } from '../common/abracadabra.common.constants';
+
 export const ARBITRUM_GLP_CAULDRONS = [
   '0x5698135ca439f21a57bddbe8b582c62f090406d5', // GLP
 ];
@@ -11,8 +13,6 @@ export const CURVE_MIM_3POOL_FARM = '0x839de324a1ab773f76a53900d70ac1b913d2b387'
 export const M_SPELL_ADDRESS = '0x1df188958a8674b5177f77667b8d173c3cdd9e51';
 export const S_SPELL_ADDRESS = '0xf7428ffcb2581a2804998efbb036a43255c8a8d3';
 export const SPELL_ADDRESS = '0xf7428ffcb2581a2804998efbb036a43255c8a8d3';
-
-export const HOURS_PER_YEAR = 8760;
 
 export const MAGIC_GLP_ADDRESS = '0x85667409a723684fe1e57dd1abde8d88c2f54214';
 export const GLP_REWARD_TRACKER_ADDRESSES = [