+++ src/apps/defiedge/binance-smart-chain/defiedge.strategy.token-fetcher.ts
@@ -1,0 +1,6 @@
+import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
+
+import { DefiedgeStrategyTokenFetcher } from '../common/defiedge.strategy.token-fetcher';
+
+@PositionTemplate()
+export class BinanceSmartChainDefiedgeStrategyTokenFetcher extends DefiedgeStrategyTokenFetcher {}