+++ src/apps/kyberswap-classic/kyberswap-classic.module.ts
@@ -9,14 +9,14 @@ import { AvalancheKyberSwapClassicPoolTokenFetcher } from './avalanche/kyberswap
 import { BinanceSmartChainKyberSwapClassicFarmContractPositionFetcher } from './binance-smart-chain/kyberswap-classic.farm.contract-position-fetcher';
 import { BinanceSmartChainKyberSwapClassicPoolTokenFetcher } from './binance-smart-chain/kyberswap-classic.pool.token-fetcher';
 import { KyberswapClassicContractFactory } from './contracts';
-import { CronosKyberSwapClassicPoolTokenFetcher } from './cronos/kyberswap-classic.pool.token-fetcher';
 import { EthereumKyberSwapClassicFarmContractPositionFetcher } from './ethereum/kyberswap-classic.farm.contract-position-fetcher';
 import { EthereumKyberSwapClassicPoolTokenFetcher } from './ethereum/kyberswap-classic.pool.token-fetcher';
 import { FantomKyberSwapClassicPoolTokenFetcher } from './fantom/kyberswap-classic.pool.token-fetcher';
 import { OptimismKyberSwapClassicPoolTokenFetcher } from './optimism/kyberswap-classic.pool.token-fetcher';
+import { PolygonKyberSwapDmmClassicPoolTokenFetcher } from './polygon/kyberswap-classic.dmm-pool.token-fetcher';
 import { PolygonKyberSwapClassicFarmContractPositionFetcher } from './polygon/kyberswap-classic.farm.contract-position-fetcher';
+import { PolygonKyberSwapKsClassicPoolTokenFetcher } from './polygon/kyberswap-classic.ks-pool.token-fetcher';
 import { PolygonKyberSwapClassicLegacyFarmContractPositionFetcher } from './polygon/kyberswap-classic.legacy-farm.contract-position-fetcher';
-import { PolygonKyberSwapClassicPoolTokenFetcher } from './polygon/kyberswap-classic.pool.token-fetcher';
 
 @Module({
   providers: [
@@ -24,12 +24,12 @@ import { PolygonKyberSwapClassicPoolTokenFetcher } from './polygon/kyberswap-cla
     UniswapV2ContractFactory,
     EthereumKyberSwapClassicPoolTokenFetcher,
     EthereumKyberSwapClassicFarmContractPositionFetcher,
-    PolygonKyberSwapClassicPoolTokenFetcher,
+    PolygonKyberSwapDmmClassicPoolTokenFetcher,
+    PolygonKyberSwapKsClassicPoolTokenFetcher,
     PolygonKyberSwapClassicFarmContractPositionFetcher,
     PolygonKyberSwapClassicLegacyFarmContractPositionFetcher,
     AvalancheKyberSwapClassicPoolTokenFetcher,
     AvalancheKyberSwapClassicFarmContractPositionFetcher,
-    CronosKyberSwapClassicPoolTokenFetcher,
     FantomKyberSwapClassicPoolTokenFetcher,
     ArbitrumKyberSwapClassicPoolTokenFetcher,
     BinanceSmartChainKyberSwapClassicFarmContractPositionFetcher,