+++ src/apps/lyra-newport/ethereum/lyra-newport.staking.contract-position-fetcher.ts
@@ -42,6 +42,10 @@ export class EthereumLyraNewportStakingContractPositionFetcher extends SingleSta
     return contract.rewardRate();
   }
 
+  async getIsActive({ contract }: GetDataPropsParams<LyraLpStaking>) {
+    return (await contract.rewardRate()).gt(0);
+  }
+
   getStakedTokenBalance({ contract, address }: GetTokenBalancesParams<LyraLpStaking, SingleStakingFarmDataProps>) {
     return contract.balanceOf(address);
   }