+++ src/apps/plutus/arbitrum/plutus.farm-pls-dpx-v2.contract-position-fetcher.ts
@@ -2,7 +2,11 @@ import { Inject } from '@nestjs/common';
 
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
 import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
-import { GetDataPropsParams, GetTokenBalancesParams } from '~position/template/contract-position.template.types';
+import {
+  GetDataPropsParams,
+  GetDisplayPropsParams,
+  GetTokenBalancesParams,
+} from '~position/template/contract-position.template.types';
 import {
   SingleStakingFarmDataProps,
   SingleStakingFarmDefinition,
@@ -11,9 +15,17 @@ import {
 
 import { PlutusContractFactory, PlutusFarmPlsDpxV2 } from '../contracts';
 
+export type PlutusFarmDefinition = SingleStakingFarmDefinition & {
+  label: string;
+};
+
 @PositionTemplate()
-export class ArbitrumPlutusFarmPlsDpxV2ContractPositionFetcher extends SingleStakingFarmTemplateContractPositionFetcher<PlutusFarmPlsDpxV2> {
-  groupLabel = 'plsDPX Farm V2';
+export class ArbitrumPlutusFarmPlsDpxV2ContractPositionFetcher extends SingleStakingFarmTemplateContractPositionFetcher<
+  PlutusFarmPlsDpxV2,
+  SingleStakingFarmDataProps,
+  PlutusFarmDefinition
+> {
+  groupLabel = 'Farms';
 
   constructor(
     @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
@@ -26,11 +38,12 @@ export class ArbitrumPlutusFarmPlsDpxV2ContractPositionFetcher extends SingleSta
     return this.contractFactory.plutusFarmPlsDpxV2({ address, network: this.network });
   }
 
-  async getFarmDefinitions(): Promise<SingleStakingFarmDefinition[]> {
+  async getFarmDefinitions(): Promise<PlutusFarmDefinition[]> {
     return [
       {
         address: '0x75c143460f6e3e22f439dff947e25c9ccb72d2e8',
         stakedTokenAddress: '0xf236ea74b515ef96a9898f5a4ed4aa591f253ce1',
+        label: 'plsDPX-v2',
         rewardTokenAddresses: [
           '0x51318b7d00db7acc4026c88c3952b66278b6a67f', // PLS
           '0xf236ea74b515ef96a9898f5a4ed4aa591f253ce1', // plsDPX
@@ -53,6 +66,29 @@ export class ArbitrumPlutusFarmPlsDpxV2ContractPositionFetcher extends SingleSta
     const duration = Date.now() / 1000 - lastRewardSecond;
     const dpxEmissions = Number(emissions.pendingDpxLessFee_) / duration;
     return [emissions.pls_, emissions.plsDpx_, emissions.plsJones_, dpxEmissions];
+  }
+
+  async getIsActive({
+    contract,
+  }: GetDataPropsParams<PlutusFarmPlsDpxV2, SingleStakingFarmDataProps, PlutusFarmDefinition>): Promise<boolean> {
+    const rewardsDistro = await contract.REWARDS_DISTRO();
+    const rewardsDistroContract = this.contractFactory.plutusRewardsDistroPlsDpxV2({
+      address: rewardsDistro,
+      network: this.network,
+    });
+
+    const emissions = await rewardsDistroContract.getEmissions();
+    const lastRewardSecond = await contract.lastRewardSecond();
+    const duration = Date.now() / 1000 - lastRewardSecond;
+    const dpxEmissions = Number(emissions.pendingDpxLessFee_) / duration;
+
+    return emissions.pls_.gt(0) || emissions.plsDpx_.gt(0) || emissions.plsJones_.gt(0) || dpxEmissions > 0;
+  }
+
+  async getLabel({
+    definition,
+  }: GetDisplayPropsParams<PlutusFarmPlsDpxV2, SingleStakingFarmDataProps, PlutusFarmDefinition>) {
+    return definition.label;
   }
 
   async getStakedTokenBalance({ contract, address }: GetTokenBalancesParams<PlutusFarmPlsDpxV2>) {