+++ src/apps/silo-finance/common/silo-finance.d-token.token-fetcher.ts
@@ -1,13 +1,12 @@
 import { Inject } from '@nestjs/common';
-import _ from 'lodash';
 
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
-import { drillBalance } from '~app-toolkit/helpers/drill-balance.helper';
-import { AppTokenPositionBalance, RawTokenBalance } from '~position/position-balance.interface';
 import { AppTokenTemplatePositionFetcher } from '~position/template/app-token.template.position-fetcher';
 import {
+  DefaultAppTokenDataProps,
   GetAddressesParams,
   GetDataPropsParams,
+  GetPricePerShareParams,
   GetUnderlyingTokensParams,
 } from '~position/template/app-token.template.types';
 
@@ -15,10 +14,7 @@ import { SiloFinanceContractFactory, SiloMarketAsset } from '../contracts';
 
 import { SiloFinanceDefinitionResolver } from './silo-finance.definition-resolver';
 
-export type DTokenDataProps = {
-  liquidity: number;
-  reserves: number[];
-  apy: number;
+export type DTokenDataProps = DefaultAppTokenDataProps & {
   siloAddress: string;
 };
 
@@ -32,10 +28,6 @@ export abstract class SiloFinanceDTokenTokenFetcher extends AppTokenTemplatePosi
   DTokenDataProps,
   DTokenDefinition
 > {
-  isExcludedFromBalances = true;
-  isExcludedFromExplore = true;
-  isExcludedFromTvl = true;
-
   isDebt = true;
 
   abstract siloLensAddress: string;
@@ -54,24 +46,15 @@ export abstract class SiloFinanceDTokenTokenFetcher extends AppTokenTemplatePosi
   }
 
   async getDefinitions(): Promise<DTokenDefinition[]> {
-    const markets = await this.siloDefinitionResolver.getSiloDefinition(this.network);
+    const markets = await this.siloDefinitionResolver.getSiloDefinitions(this.network);
     if (!markets) return [];
 
-    const dTokenDefinition = markets
-      .map(market => {
-        const siloAddress = market.siloAddress;
-        const dTokenAddresses = market.marketAssets.map(x => x.dToken);
-
-        return dTokenAddresses.map(address => {
-          return {
-            address,
-            siloAddress,
-          };
-        });
-      })
-      .flat();
-
-    return dTokenDefinition;
+    return markets.flatMap(market =>
+      market.marketAssets.map(marketAsset => ({
+        siloAddress: market.siloAddress,
+        address: marketAsset.dToken,
+      })),
+    );
   }
 
   async getAddresses({ definitions }: GetAddressesParams<DTokenDefinition>): Promise<string[]> {
@@ -82,70 +65,25 @@ export abstract class SiloFinanceDTokenTokenFetcher extends AppTokenTemplatePosi
     return [{ address: await contract.asset(), network: this.network }];
   }
 
-  async getPricePerShare() {
-    return [1];
+  async getPricePerShare({
+    multicall,
+    definition,
+    appToken,
+  }: GetPricePerShareParams<SiloMarketAsset, DTokenDataProps, DTokenDefinition>): Promise<number[]> {
+    const siloLensContract = this.contractFactory.siloLens({ address: this.siloLensAddress, network: this.network });
+    const reserveRaw = await multicall
+      .wrap(siloLensContract)
+      .totalBorrowAmountWithInterest(definition.siloAddress, appToken.tokens[0].address);
+    if (reserveRaw.isZero()) return [0];
+
+    const reserve = Number(reserveRaw) / 10 ** appToken.tokens[0].decimals;
+    const pricePerShare = reserve / Number(appToken.supply);
+    return [pricePerShare];
   }
 
   async getDataProps(params: GetDataPropsParams<SiloMarketAsset, DTokenDataProps, DTokenDefinition>) {
     const defaultDataProps = await super.getDataProps(params);
     const siloAddress = params.definition.siloAddress;
     return { ...defaultDataProps, siloAddress };
-  }
-
-  async getRawBalances(address: string): Promise<RawTokenBalance[]> {
-    const multicall = this.appToolkit.getMulticall(this.network);
-
-    const appTokens = await this.appToolkit.getAppTokenPositions<DTokenDataProps>({
-      appId: this.appId,
-      network: this.network,
-      groupIds: [this.groupId],
-    });
-
-    const siloLensContract = this.contractFactory.siloLens({ address: this.siloLensAddress, network: this.network });
-
-    return (
-      await Promise.all(
-        appTokens.map(async appToken => {
-          const balanceRaw = await this.getBalancePerToken({ multicall, address, appToken });
-          const totalBorrowWithInterest = await multicall
-            .wrap(siloLensContract)
-            .totalBorrowAmountWithInterest(appToken.dataProps.siloAddress, appToken.tokens[0].address);
-
-          return [
-            {
-              key: this.appToolkit.getPositionKey(appToken),
-              balance: balanceRaw.toString(),
-            },
-            {
-              key: `${this.appToolkit.getPositionKey(appToken)}-underlying`,
-              balance: totalBorrowWithInterest.toString(),
-            },
-          ];
-        }),
-      )
-    ).flat();
-  }
-
-  async drillRawBalances(balances: RawTokenBalance[]): Promise<AppTokenPositionBalance<DTokenDataProps>[]> {
-    const appTokens = await this.getPositionsForBalances();
-
-    const appTokenBalances = appTokens.map(token => {
-      const tokenBalance = balances.find(b => b.key === this.appToolkit.getPositionKey(token));
-      const underlyingTokenBalance = balances.find(
-        b => b.key === `${this.appToolkit.getPositionKey(token)}-underlying`,
-      );
-
-      if (!tokenBalance || !underlyingTokenBalance) return null;
-
-      const result = drillBalance<typeof token, DTokenDataProps>(token, tokenBalance.balance, {
-        isDebt: this.isDebt,
-      });
-
-      const underlyingToken = drillBalance<typeof token, DTokenDataProps>(appTokens[0], underlyingTokenBalance.balance);
-
-      return { ...result, tokens: [underlyingToken] };
-    });
-
-    return _.compact(appTokenBalances);
   }
 }