+++ src/apps/sonne/optimism/sonne.borrow.contract-position-fetcher.ts
@@ -3,22 +3,27 @@ import { Inject } from '@nestjs/common';
 import { APP_TOOLKIT, IAppToolkit } from '~app-toolkit/app-toolkit.interface';
 import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
 import {
+  buildDollarDisplayItem,
+  buildPercentageDisplayItem,
+  buildStringDisplayItem,
+} from '~app-toolkit/helpers/presentation/display-item.present';
+import {
   CompoundBorrowContractPositionFetcher,
+  CompoundBorrowTokenDataProps,
   GetMarketsParams,
 } from '~apps/compound/common/compound.borrow.contract-position-fetcher';
 import { isMulticallUnderlyingError } from '~multicall/multicall.ethers';
+import { StatsItem } from '~position/display.interface';
 import {
   GetDataPropsParams,
+  GetDisplayPropsParams,
   GetTokenBalancesParams,
   GetTokenDefinitionsParams,
 } from '~position/template/contract-position.template.types';
 
 import { SonneComptroller, SonneContractFactory, SonneSoToken } from '../contracts';
 
-export type SonneBorrowTokenDataProps = {
-  apy: number;
-  liquidity: number;
-  isActive: boolean;
+export type SonneBorrowTokenDataProps = CompoundBorrowTokenDataProps & {
   collateralFactor: number;
 };
 
@@ -80,13 +85,14 @@ export class OptimismSonneBorrowContractPositionFetcher extends CompoundBorrowCo
   async getDataProps(
     params: GetDataPropsParams<SonneSoToken, SonneBorrowTokenDataProps>,
   ): Promise<SonneBorrowTokenDataProps> {
+    const defaultDataProps = await super.getDataProps(params);
+
     const [underlyingToken] = params.contractPosition.tokens;
 
-    const [decimals, supplyRaw, pricePerShare, apy, cashRaw] = await Promise.all([
+    const [decimals, supplyRaw, pricePerShare, cashRaw] = await Promise.all([
       this.getCTokenDecimals(params),
       this.getCTokenSupply(params),
       this.getPricePerShare(params),
-      this.getApy(params),
       this.getCash(params).catch(e => {
         if (isMulticallUnderlyingError(e)) return 0;
         throw e;
@@ -113,10 +119,24 @@ export class OptimismSonneBorrowContractPositionFetcher extends CompoundBorrowCo
       borrowedPositionliquidity > underlyingLiquidity ? 0 : underlyingLiquidity - borrowedPositionliquidity;
 
     return {
+      ...defaultDataProps,
       liquidity: -borrowLiquidity,
       isActive: Boolean(borrowLiquidity > 0),
-      apy,
       collateralFactor,
     };
+  }
+
+  async getStatsItems({
+    contractPosition,
+  }: GetDisplayPropsParams<SonneSoToken, SonneBorrowTokenDataProps>): Promise<StatsItem[] | undefined> {
+    const { apy, liquidity, exchangeRate, symbol, collateralFactor } = contractPosition.dataProps;
+    const exchangeRateFormatted = `${(1 / exchangeRate).toFixed(3)} ${symbol} = 1 ${contractPosition.tokens[0].symbol}`;
+
+    return [
+      { label: 'APY', value: buildPercentageDisplayItem(apy) },
+      { label: 'Liquidity', value: buildDollarDisplayItem(liquidity) },
+      { label: 'Exchange Rate', value: buildStringDisplayItem(exchangeRateFormatted) },
+      { label: 'Collateral Factor', value: buildPercentageDisplayItem(collateralFactor) },
+    ];
   }
 }