+++ CHANGELOG.md
@@ -4,6 +4,218 @@
 
 [1]: https://www.npmjs.com/package/@zapper-fi/studio?activeTab=versions
 
+## [0.438.1](https://github.com/Zapper-fi/studio/compare/v0.438.0...v0.438.1) (2023-04-07)
+
+
+### Bug Fixes
+
+* **studio:** Fix zero totals case ([#2546](https://github.com/Zapper-fi/studio/issues/2546)) ([a06ac7b](https://github.com/Zapper-fi/studio/commit/a06ac7bf5f00327a7e3ec6d9c6aafc7ff6d090bc))
+
+## [0.438.0](https://github.com/Zapper-fi/studio/compare/v0.437.1...v0.438.0) (2023-04-07)
+
+
+### Features
+
+* **studio:** Fix Silo Finance everything ([#2544](https://github.com/Zapper-fi/studio/issues/2544)) ([6978ec5](https://github.com/Zapper-fi/studio/commit/6978ec5e800398d2e2f8f8c72b57cfab0ffa6c9a))
+
+## [0.437.1](https://github.com/Zapper-fi/studio/compare/v0.437.0...v0.437.1) (2023-04-07)
+
+
+### Bug Fixes
+
+* **studio:** Revamping KyberSwap classic pools+farms ([7a3f5aa](https://github.com/Zapper-fi/studio/commit/7a3f5aae5915f66a20c63fec6c34dab67d8eec12))
+* **studio:** Revamping KyberSwap classic pools+farms ([#2542](https://github.com/Zapper-fi/studio/issues/2542)) ([7a3f5aa](https://github.com/Zapper-fi/studio/commit/7a3f5aae5915f66a20c63fec6c34dab67d8eec12))
+
+## [0.437.0](https://github.com/Zapper-fi/studio/compare/v0.436.0...v0.437.0) (2023-04-07)
+
+
+### Features
+
+* **abracadabra:** Add magicGLP for Avalanche and protocol debt cauldron ([#2539](https://github.com/Zapper-fi/studio/issues/2539)) ([40f9ad1](https://github.com/Zapper-fi/studio/commit/40f9ad1a718b86be61c0a24d414f244ad3b4ae81))
+
+## [0.436.0](https://github.com/Zapper-fi/studio/compare/v0.435.0...v0.436.0) (2023-04-07)
+
+
+### Features
+
+* **main:** add reserve protocol app ([#2508](https://github.com/Zapper-fi/studio/issues/2508)) ([e264165](https://github.com/Zapper-fi/studio/commit/e264165847c42f8966aa4f04a40e414791d575bb))
+
+## [0.435.0](https://github.com/Zapper-fi/studio/compare/v0.434.1...v0.435.0) (2023-04-07)
+
+
+### Features
+
+* **studio:** Make resolving is active abstract ([#2534](https://github.com/Zapper-fi/studio/issues/2534)) ([8433ec1](https://github.com/Zapper-fi/studio/commit/8433ec1742b06e5895b1bdcb00500dc10a8683d7))
+
+
+### Bug Fixes
+
+* **studio:** Update KyberSwap SDK to latest version ([#2533](https://github.com/Zapper-fi/studio/issues/2533)) ([11354f2](https://github.com/Zapper-fi/studio/commit/11354f2723674692522d70f9a19504f4ca963b43))
+
+## [0.434.1](https://github.com/Zapper-fi/studio/compare/v0.434.0...v0.434.1) (2023-04-06)
+
+
+### Bug Fixes
+
+* **plutus:** Fix plsRdnt token order ([#2535](https://github.com/Zapper-fi/studio/issues/2535)) ([082b73d](https://github.com/Zapper-fi/studio/commit/082b73d6f9a43d97925592b1dfa2b2b649574e3d))
+
+## [0.434.0](https://github.com/Zapper-fi/studio/compare/v0.433.1...v0.434.0) (2023-04-06)
+
+
+### Features
+
+* **exactly:** ✨ add claimable rewards ([#2531](https://github.com/Zapper-fi/studio/issues/2531)) ([cd99062](https://github.com/Zapper-fi/studio/commit/cd99062ea8bf692f82d073f3da687490889fd643))
+
+## [0.433.1](https://github.com/Zapper-fi/studio/compare/v0.433.0...v0.433.1) (2023-04-06)
+
+
+### Bug Fixes
+
+* **apps:** Remove app integrations on Cronos and Evmos ([#2529](https://github.com/Zapper-fi/studio/issues/2529)) ([b0af1e7](https://github.com/Zapper-fi/studio/commit/b0af1e745470e7c2ecd2ddaf8a0d67606bf29035))
+
+## [0.433.0](https://github.com/Zapper-fi/studio/compare/v0.432.0...v0.433.0) (2023-04-06)
+
+
+### Features
+
+* **uniswap-v3:** Show min/max ranges and fix sig figs on range ([#2528](https://github.com/Zapper-fi/studio/issues/2528)) ([a341f75](https://github.com/Zapper-fi/studio/commit/a341f75a0bca624357c621d7c06845d668414204))
+
+
+### Bug Fixes
+
+* **lyra-avalon:** Fix option decimals on Arbitrum ([#2525](https://github.com/Zapper-fi/studio/issues/2525)) ([2982616](https://github.com/Zapper-fi/studio/commit/2982616c6a7de827de9602662285a21415ad48c0))
+* **thales:** Linting ([#2526](https://github.com/Zapper-fi/studio/issues/2526)) ([6a4a9b3](https://github.com/Zapper-fi/studio/commit/6a4a9b3143dbb00d3c6783e43b836b35e182cd3d))
+
+## [0.432.0](https://github.com/Zapper-fi/studio/compare/v0.431.0...v0.432.0) (2023-04-05)
+
+
+### Features
+
+* **thales:** Arbitrum Vaults ([#2517](https://github.com/Zapper-fi/studio/issues/2517)) ([cd93a44](https://github.com/Zapper-fi/studio/commit/cd93a44f4c08faf519cb70a17d72c391b7365287))
+
+
+### Bug Fixes
+
+* **convex:** Resolve underlying token from stash token wrappers ([#2523](https://github.com/Zapper-fi/studio/issues/2523)) ([25282d4](https://github.com/Zapper-fi/studio/commit/25282d4cfa2fdf868e8bc281b49094a1b1274686))
+* **plutus:** Fix farms ([#2522](https://github.com/Zapper-fi/studio/issues/2522)) ([62282ac](https://github.com/Zapper-fi/studio/commit/62282ac721bdcb68fea0f1c955db5331afe77ef7))
+
+## [0.431.0](https://github.com/Zapper-fi/studio/compare/v0.430.1...v0.431.0) (2023-04-05)
+
+
+### Features
+
+* **studio:** Pass through isDebt as a data prop ([#2518](https://github.com/Zapper-fi/studio/issues/2518)) ([f45895e](https://github.com/Zapper-fi/studio/commit/f45895ea49b5d454e43180ce5d069a10045bcc25))
+* **studio:** Use address in getPositionsForBalances to filter down where necessary ([#2520](https://github.com/Zapper-fi/studio/issues/2520)) ([70c8bab](https://github.com/Zapper-fi/studio/commit/70c8bab70976704aeb30d6062f4ff632197c089f))
+
+## [0.430.1](https://github.com/Zapper-fi/studio/compare/v0.430.0...v0.430.1) (2023-04-05)
+
+
+### Bug Fixes
+
+* **lyra-avalon:** Remove staked Lyra claimables position on Arbitrum ([#2515](https://github.com/Zapper-fi/studio/issues/2515)) ([03a4c68](https://github.com/Zapper-fi/studio/commit/03a4c68d4e2e76a3957a2f2e091369347201e55a))
+
+## [0.430.0](https://github.com/Zapper-fi/studio/compare/v0.429.0...v0.430.0) (2023-04-04)
+
+
+### Features
+
+* **silo-finance:** Add Incentive position on Ethereum ([#2513](https://github.com/Zapper-fi/studio/issues/2513)) ([f0e8a1e](https://github.com/Zapper-fi/studio/commit/f0e8a1e1796899d369fa6149d01eb273fd7c2b0f))
+
+## [0.429.0](https://github.com/Zapper-fi/studio/compare/v0.428.0...v0.429.0) (2023-04-04)
+
+
+### Features
+
+* **lyra-avalon:** Add staked Lyra and Options on Arbitrum ([#2511](https://github.com/Zapper-fi/studio/issues/2511)) ([338d96f](https://github.com/Zapper-fi/studio/commit/338d96f00dadb45a77bfbb42bd387c82f7106ac4))
+
+## [0.428.0](https://github.com/Zapper-fi/studio/compare/v0.427.0...v0.428.0) (2023-04-04)
+
+
+### Features
+
+* **compound:** Add exchange rate to supply and borrow positions ([#2509](https://github.com/Zapper-fi/studio/issues/2509)) ([4fbfeb1](https://github.com/Zapper-fi/studio/commit/4fbfeb1134cf43707318aa92a6240ad7818984a0))
+
+
+### Bug Fixes
+
+* **exactly:** 🐛 variable borrow label ([#2505](https://github.com/Zapper-fi/studio/issues/2505)) ([f4ec7d7](https://github.com/Zapper-fi/studio/commit/f4ec7d732b9745f430a4afa27de4397225d75032))
+* **studio:** Fixing balance builder for UniV3 positions on Celo ([#2510](https://github.com/Zapper-fi/studio/issues/2510)) ([1730aad](https://github.com/Zapper-fi/studio/commit/1730aadcb747e79056140b5f47c91c16797a29a0))
+
+## [0.427.0](https://github.com/Zapper-fi/studio/compare/v0.426.0...v0.427.0) (2023-04-03)
+
+
+### Features
+
+* **silo-finance:** claimable SILO on Arbitrum ([#2492](https://github.com/Zapper-fi/studio/issues/2492)) ([9348fd4](https://github.com/Zapper-fi/studio/commit/9348fd475402fd0eda8362c531439ab2c372690d))
+
+## [0.426.0](https://github.com/Zapper-fi/studio/compare/v0.425.2...v0.426.0) (2023-04-03)
+
+
+### Features
+
+* **studio:** Update logic for BeethovenX token supplies ([#2502](https://github.com/Zapper-fi/studio/issues/2502)) ([3282728](https://github.com/Zapper-fi/studio/commit/3282728f68be03edb2b163d5dfe5a6843f7f5a89))
+
+## [0.425.2](https://github.com/Zapper-fi/studio/compare/v0.425.1...v0.425.2) (2023-04-03)
+
+
+### Bug Fixes
+
+* **yield-protocol:** Fix failing positions ([#2500](https://github.com/Zapper-fi/studio/issues/2500)) ([7e57fa5](https://github.com/Zapper-fi/studio/commit/7e57fa58e6d506a78285b35e32926f9a7ed7d0cc))
+
+## [0.425.1](https://github.com/Zapper-fi/studio/compare/v0.425.0...v0.425.1) (2023-04-03)
+
+
+### Bug Fixes
+
+* **pool-together-v4:** Fix ticket symbol ([#2498](https://github.com/Zapper-fi/studio/issues/2498)) ([c41d92d](https://github.com/Zapper-fi/studio/commit/c41d92deeaffcdc64cf2112774c337bdcffa3d60))
+
+## [0.425.0](https://github.com/Zapper-fi/studio/compare/v0.424.0...v0.425.0) (2023-04-03)
+
+
+### Features
+
+* **dopex:** Filter out expired options ([#2496](https://github.com/Zapper-fi/studio/issues/2496)) ([d478aa9](https://github.com/Zapper-fi/studio/commit/d478aa920bc03f45c9f94d25045284c40031118b))
+
+## [0.424.0](https://github.com/Zapper-fi/studio/compare/v0.423.0...v0.424.0) (2023-04-03)
+
+
+### Features
+
+* **idle:** Filter out paused vault tokens that crash the fetcher ([#2494](https://github.com/Zapper-fi/studio/issues/2494)) ([7500f77](https://github.com/Zapper-fi/studio/commit/7500f77e95f71eaebb538cf0aab59a8b6e0addc8))
+
+## [0.423.0](https://github.com/Zapper-fi/studio/compare/v0.422.0...v0.423.0) (2023-04-03)
+
+
+### Features
+
+* **ramses:** Fix pools, use on-chain requests only ([#2491](https://github.com/Zapper-fi/studio/issues/2491)) ([dc1a899](https://github.com/Zapper-fi/studio/commit/dc1a899455a1d2a6f3f5ccf1d3591f2123689008))
+
+## [0.422.0](https://github.com/Zapper-fi/studio/compare/v0.421.0...v0.422.0) (2023-04-01)
+
+
+### Features
+
+* **defiedge:** bsc and polygon farming ([#2485](https://github.com/Zapper-fi/studio/issues/2485)) ([fff05e1](https://github.com/Zapper-fi/studio/commit/fff05e199bfe88e7174861236ea271492f9bf037))
+
+## [0.421.0](https://github.com/Zapper-fi/studio/compare/v0.420.0...v0.421.0) (2023-04-01)
+
+
+### Features
+
+* **synthetix:** perp refactor ([#2486](https://github.com/Zapper-fi/studio/issues/2486)) ([08c2454](https://github.com/Zapper-fi/studio/commit/08c2454f7e588ce130095889ce2a8217d8515832))
+
+
+### Bug Fixes
+
+* **llama-airforce:** Add new uCRV pounder and distributor ([#2487](https://github.com/Zapper-fi/studio/issues/2487)) ([e85021a](https://github.com/Zapper-fi/studio/commit/e85021a3f975906319357cebfd47d1772c187be2))
+
+## [0.420.0](https://github.com/Zapper-fi/studio/compare/v0.419.0...v0.420.0) (2023-03-30)
+
+
+### Features
+
+* **dolomite:** Added Dolomite to Zapper ⛰️  ([#2477](https://github.com/Zapper-fi/studio/issues/2477)) ([d8429e4](https://github.com/Zapper-fi/studio/commit/d8429e44de5400ecf70109a24de8905d082276b5))
+
 ## [0.419.0](https://github.com/Zapper-fi/studio/compare/v0.418.3...v0.419.0) (2023-03-30)